nie jest dobrze :(

+++ src/main/java/org/example/audiobook.java
@@ -1,18 +1,21 @@
 package org.example;
 
-public class audiobook extends item{
-    private float size;
+public class audiobook extends item {
+    private int size;
 
     public float getSize() {
         return size;
     }
 
-    public void setSize(float size) {
+    public void setSize(int size) {
         this.size = size;
     }
 
-    public audiobook(String name, String author, int publictionDate, int length, literaryGenres genre, float size) {
+    public audiobook(String name, String author, int publictionDate, int length, literaryGenres genre, int size) {
         super(name, author, publictionDate, length, genre);
         this.size = size;
     }
 }
+
+
+
